fix: read package.json version at runtime instead of bundle time
