+++ packages/zudoku/src/cli/dev/handler.ts
@@ -1,8 +1,9 @@
 import path from "node:path";
-import packageJson from "../../../package.json" with { type: "json" };
+import { fileURLToPath } from "node:url";
 import { joinUrl } from "../../lib/util/joinUrl.js";
 import { DevServer } from "../../vite/dev-server.js";
 import { printDiagnosticsToConsole } from "../common/output.js";
+import { getPackageJson } from "../common/package-json.js";
 
 export interface Arguments {
   dir: string;
@@ -12,6 +13,9 @@ export interface Arguments {
 }
 
 export async function dev(argv: Arguments) {
+  const packageJson = getPackageJson(
+    fileURLToPath(import.meta.resolve("zudoku/package.json")),
+  );
   process.env.NODE_ENV = "development";
   const dir = path.resolve(process.cwd(), argv.dir);
   const server = new DevServer({
